chore: refactor omitUnreachableAnswers into its own file
